requirements.txt
requests
boto3
python-telegram-bot